Time: 3 ms (54.53%), Space: 13.4 MB (16.26%) - LeetHub

+++ 2692-take-gifts-from-the-richest-pile/2692-take-gifts-from-the-richest-pile.cpp
@@ -7,7 +7,7 @@ public:
             maxheap.pop();
             maxheap.push(floor(sqrt(n)));
         }
-        double total = 0;
+        long int total = 0;
         while (!maxheap.empty()) {
         total += maxheap.top();
         maxheap.pop();
